Jupyter Notebook of Final Machine Learning Project
This code is part of the group project for my Machine Learning class at the University of St. Thomas. The part of the code I had the largest hand in was the Radial Base Function Kernel Support Vector Machine Model. I also reviewed all the other models and processes in the file.